requirements.txt
pymupdf>=1.24.0
langchain-text-splitters>=0.2.0
sentence-transformers>=3.0.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
google-genai>=1.0.0
streamlit>=1.38.0
torch>=2.0.0
